Added Peasant halfway done im kinda stuck
How do i create an ability to give to Peasant on creation of the Character, it seems ok given that a Character Archetype should be static but its hard to create it that way with the current structure.

+++ hexarena/Scripts/hexarena/GameLogic/Map/Tile.cs
@@ -20,7 +20,7 @@ namespace GameLogic{
 			Neighbours = new List<ITile>();
 		}
 		
-		public override void TakeDamage(int damage){
+		public override void TakeDamage(short damage){
 			if(CharacterOnTile == null){return;}
 			else{
 				CharacterOnTile.Health -= damage;
