+++ hexarena/Scripts/hexarena/Interfaces/ICharacter.cs
@@ -9,7 +9,7 @@ namespace Interfaces
 	public interface ICharacter : ITargetable
 	{	
 		double StepEnergyCost { get;  set; }
-		public int Health { get; set; }
+		public short Health { get; set; }
 		public List<IStatusEffect> StatusEffects {get; protected set;}
 		public List<IActive> ActiveAbilities {get; protected set;} 
 		public List<IPassive> CharacterPassives {get; protected set;}  //ne znam dali ni trqbva kato imame StatusEffects List already ama mi se struva pravilno
